[WIP] Add host decoder program for `array_test`...
The following test on the host suggests that the `ArrayTest` request type works
correctly running on the host _(as opposed to on the Arduino target)_:

    $ ./decode 3502_2.dat
    Got ArrayTestRequest: 24
    data: 3, 5, 0, 2
    index: 2

The output above shows that the array `[3, 5, 0, 2]` is properly decoded by the
`nanopb` decode function.  This result suggests that the issues decoding
described in 458b06b5e34d4338482bdff6b7313d1353f97c2c are isolated to AVR
controllers.

TODO
====

 - Enable Arduino command-processor debug mode, where received requests may be
   dumped to the serial output stream for debugging purposes.