[NB] Add method to save i2c address to EEPROM
Add method to save I2C address to EEPROM, which persists I2C address after
resetting the device.  This makes it possible to connect to the device using
the serial port to set the I2C address, after which an I2C connection can be
used to interface with the device.