[NB] Add `array_demo` method.
Add `array_demo` method, which accepts an array of bytes and an index, and
returns the value stored in the array position corresponding to the index.
This method makes it possible to verify that the array is received intact on
the device.  As of this commit, arrays of size 40 have been verified to send
correctly, including those that contain a value of zero.  This fixes #1.