Add RPCBuffer.c to construct RPC-related buffers
 - Add `RPCBuffer.c` to construct RPC-related global buffers.
 - Add `RPCBuffer.ht` as [Jinja2][1] template for generating
   `RPCBuffer.h` header.
  - Header defines size of:
   - Serial interface packet.
   - i2c interface packet.
   - Command processor array buffer, which is used for decoding arrays
     received as RPC arguments.

Header generation
=================

Generating default header
-------------------------

A `RPCBuffer.h` header can be generated by using the following command:

    paver generate_rpc_buffer_header

The `generate_rpc_buffer_header` paver task is also a dependency of the
`build_firmware` task, so the header will automatically be generated
when running:

    paver build_firmware

Overriding default header
-------------------------

The header will only be generated if there is no existing file.  This
permits users to make modifications to the header.