Add option to set default for unset fields
Add option to set default if field value has not been set in calls to
`get_field_value`.

Fix bug in enum type name lookup.  Prior to this commit, name was looked
up based on value in 0-indexed contiguous value list, but this is wrong
since enum values do not need to be contiguous.  As of this commit,
names are correctly looked up based on the value in the
`values_by_number` dictionary.